+++ CHANGELOG.md
@@ -29,6 +29,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Load kernel module [nvidia-uvm](https://developer.nvidia.com/blog/unified-memory-cuda-beginners/) by default to provide Unified Virtual Memory (UVM) functionality to the CUDA driver.
 - Install [NVIDIA Persistence Daemon](https://docs.nvidia.com/deploy/driver-persistence/index.html) as a system service.
 - Install [NVIDIA Data Center GPU Manager (DCGM)](https://developer.nvidia.com/dcgm) package on all supported OSes except for aarch64 `centos7` and `alinux2`.
+- Forward SLURM_RESUME_FILE to ParallelCluster resume program.
 
 **CHANGES**
 - Upgrade Slurm to version 23.02.2.